Finish creation functionality
